app/requirements.txt
Flask
mysql-connector
flask-migrate
gunicorn==20.0.4
